+++ app.js
@@ -13,11 +13,11 @@ const expressErrorHandler = require('express-error-handler');
 // Session 미들웨어 불러오기
 const session = require('express-session');
 
-// crypto 모듈 불러오기
-const crypto = require('crypto');
-
 // Express 객체 생성
 const app = express();
+
+// Route 객체 생성
+const userRoute = require('./routes/userRoute');
 
 // 기본 속성 설정
 app.set('port', process.env.PORT || 3000);
@@ -41,15 +41,11 @@ app.use(session({
     saveUninitialized: true
 }));
 
-
 // mongoose 모듈 불러오기
 const mongoose = require('mongoose');
 
 // 데이터베이스 객체를 위한 변수 선언
 let database;
-
-// 데이터베이스 스키마 객체를 위한 변수 선언
-let userSchema;
 
 // 데이터베이스 모델 객체를 위한 변수 선언
 let userModel;
@@ -81,254 +77,26 @@ function connectDB() {
 }
 
 function createUserSchema() {
-    /*
-    * 스키마 정의
-    * password를 hashedPassword로 변경
-    * default 속성 추가
-    * salt 속성 추가
-    */
-    userSchema = mongoose.Schema({
-        id: { type: String, required: true, unique: true, 'default': ' ' },
-        name: { type: String, index: 'hashed', 'default': ' ' },
-        hashedPassword: { type: String, required: true, 'default': ' ' },
-        salt: { type: String, required: true },
-        age: { type: Number, 'default': -1 },
-        createdAt: { type: Date, index: { unique: false }, 'default': Date.now },
-        updatedAt: { type: Date, index: { unique: false }, 'default': Date.now }
-    });
+    // userSchema 모듈 불러오기
+    const userSchema = require('./database/userSchema').createSchema(mongoose);
     
-    // password를 virtual 메소드로 정의
-    userSchema
-        .virtual('password')
-        .set(function(password) {
-            this._password = password;
-            this.salt = this.makeSalt();
-            this.hashedPassword = this.encryptPassword(this._password, this.salt);
-            console.log(`virtual password call : ${this.hashedPassword}`);
-        })
-        .get(function() {
-            return this._password;
-        });
-    
-    // 스키마에 모델 인스턴스에서 사용할 수 있는 메소드 추가
-    // 비밀번호 암호화 메소드
-    userSchema.method('encryptPassword', (plainText, inSalt) => {
-        if (inSalt) {
-            return crypto.createHmac('sha1', inSalt).update(plainText).digest('hex');
-        } else {
-            return crypto.createHmac('sha1', this.salt).update(plainText).digest('hex');
-        }
-    });
-    
-    // salt 생성 메소드
-    userSchema.method('makeSalt', () => {
-        return Math.round((new Date().valueOf() * Math.random())) + '';
-    });
-    
-    // 인증 메소드 (입력된 비밀번호와 비교)
-    userSchema.method('authenticate', (plainText, inSalt, hashedPassword) => {
-        if (inSalt) {
-            console.log(`authenticate call : ${plainText} -> ${this.encryptPassword(plainText, inSalt)} : ${this.hashedPassword}`);
-            return this.encryptPassword(plainText, inSalt) === hashedPassword;
-        } else {
-            console.log(`authenticate call : ${plainText} -> ${this.encryptPassword(plainText)} : ${this.hashedPassword}`);
-            return this.encryptPassword(plainText) === hashedPassword;
-        }
-    });
-    
-    // 필수 속성에 대한 유효성 확인 (길이, 값 체크)
-    userSchema.path('id').validate((id) => {
-        return id.length;
-    }, 'id 필드 값이 없습니다');
-    
-    userSchema.path ('name').validate((name) => {
-        return name.length;
-    }, 'name 필드 값이 없습니다');
-    console.log('스키마 정의 완료');
-    
+    // userModel 정의
     userModel = mongoose.model('users3', userSchema);
-    console.log('모델 정의 완료');
+    console.log('User Model 정의 완료');
+    
+    // init 호출
+    userRoute.init(database, userSchema, userModel);
 }
 
-
-// 사용자 인증
-const authUser = function(database, id, password, callback) {
-    console.log('authUser call');
-    
-    // 아이디와 비밀번호로 사용자 검색
-    userModel.findById(id, function(err, result) {
-        if (err) {
-            callback(err, null);
-        }
-    
-        console.log(`아이디 : ${id} 비밀번호 : ${password}로 조회`);
-        console.dir(result);
-        
-        if (result.length > 0) {
-            console.log('사용자 아이디 검색 완료');
-            
-            // 비밀번호 확인
-            const user = new userModel({ id: id});
-            const authenticated = user.authenticate(password, result[0]._doc.salt, result[0]._doc.hashedPassword);
-            
-            if (authenticated) {
-                console.log('비밀번호 일치');
-                callback(null, result);
-            } else {
-                console.log('비밀번호 불일치');
-                callback(null, null);
-            }
-        } else {
-            console.log('사용자 검색 불가');
-            callback(null, null);
-        }
-    });
-};
-
-// 사용자 전체 조회
-app.post('/process/userlist',(req, res) => {
-    console.log('/process/userlist call');
-    
-    // 데이터 베이스 객체가 초기화된 경우, 모델 객체의 findAll 메소드 호출
-    if (database) {
-        // 1. 모든 사용자 검색
-        userModel.findAll((err, result) => {
-            // 에러 발생 시 클라이언트로 에러 응답 전송
-            if (err) {
-                console.log(`사용자 전체 조회 중 에러 발생 : ${err.stack}`);
-                
-                res.writeHead('200', { "Content-Type": "text/html;charset=utf8"});
-                res.write(`<h2>사용자 전체 조회 중 에러 발생</h2>`);
-                res.write(`<p>${err.stack}</p>`);
-                res.end();
-            }
-            
-            // 결과 있으면 목록을 응답 전송
-            if (result) {
-                console.dir(result);
-    
-                res.writeHead('200', { "Content-Type": "text/html;charset=utf8"});
-                res.write(`<h2>사용자 전체 목록</h2>`);
-                res.write(`<div><ul>`);
-                
-                for (let i = 0; i < result.length; i++) {
-                    const curId = result[i]._doc.id;
-                    const curName = result[i]._doc.name;
-                    res.write(`    <li>#${i} : ${curId}, ${curName}</li>`);
-                }
-    
-                res.write(`</ul></div>`);
-                res.end();
-            // 결과 객체 없으면 응답 전송
-            } else {
-                res.writeHead('200', { "Content-Type": "text/html;charset=utf8"});
-                res.write(`<h2>사용자 전체 조회 실패</h2>`);
-                res.end();
-            }
-        });
-    }
-})
-
-
-
-// 라우터 객체 참조
+// 라우터 객체 생성
 const router = express.Router();
-
-// 로그인 라우팅 함수 - 데이터베이스 정보와 비교
-app.post('/process/login', function(req, res) {
-    console.log('/process/login call');
-    
-    const paramId = req.body.id || req.query.id;
-    const paramPassword = req.body.password || req.query.password;
-    
-    if (database) {
-        authUser(database, paramId, paramPassword, function(err, docs) {
-            if (err) throw err;
-            
-            if (docs) {
-                console.dir(docs);
-                const userName = docs[0].name;
-                res.writeHead('200', { 'Content-Type': 'text/html;charset=utf8'});
-                res.write(`<h1>로그인 성공</h1>`);
-                res.write(`<div><p>사용자 아이디 : ` + paramId + `</p></div>`);
-                res.write(`<div><p>사용자 이름 : ` + userName + `</p></div>`);
-                res.write(`<br><br><a href="/public/login.html">다시 로그인하기</a>`);
-                res.end();
-            } else {
-                res.writeHead('200', { 'Content-Type': 'text/html;charset=utf8'});
-                res.write(`<h1>로그인 실패</h1>`);
-                res.write(`<div><p>아이디와 비밀번호를 다시 확인하십시오</p></div>`);
-                res.write(`<br><br><a href="/public/login.html">다시 로그인하기</a>`);
-                res.end();
-            }
-        });
-    } else {
-        res.writeHead('200', { 'Content-Type': 'text/html;charset=utf8'});
-        res.write(`<h1>데이터베이스 연결 실패</h1>`);
-        res.write(`<div><p>데이터베이스를 찾지 못했습니다</p></div>`);
-        res.end();
-    }
-});
-
-// 사용자 추가
-const addUser = function(database, id, name, password, callback) {
-    console.log('addUser call');
-    
-    const user = new userModel({
-        id: id,
-        name: name,
-        password: password
-    });
-    
-    user.save((err) => {
-        if (err) {
-            callback(err, null);
-        }
-    
-        console.log('사용자 추가 완료');
-        callback(null, user);
-    });
-};
-
-// 사용자 추가 라우팅
-router.route('/process/adduser').post((req, res) => {
-    console.log('/process/adduser call');
-    
-    const paramId = req.body.id || req.query.id;
-    const paramName = req.body.name || req.query.name;
-    const paramPassword = req.body.password || req.query.password;
-    
-    console.log('사용자 추가 요청 파라미터 : ' + paramId +', ' + paramName + ', ' + paramPassword);
-    
-    // 데이터베이스 객체가 초기화된 경우, addUser 함수 호출하여 사용자 추가
-    if (database) {
-        addUser(database, paramId, paramName, paramPassword, (err, result) => {
-            if (err) throw err;
-    
-            // 결과 객체 확인하여 추가된 데이터가 있으면 성공 응답 전송
-            if (result.insertedId !== null) {
-                console.dir(result);
-                
-                res.writeHead('200', { "Content-Type": "text/html;charset=utf8"});
-                res.write(`<h2>사용자 추가 성공</h2>`);
-                res.end();
-            } else {
-                res.writeHead('200', { "Content-Type": "text/html;charset=utf8"});
-                res.write(`<h2>사용자 추가 실패</h2>`);
-                res.end();
-            }
-        });
-        // 데이터 베이스 객체과 초기화 되지 않은 경우
-    } else {
-        res.writeHead('200', { "Content-Type": "text/html;charset=utf8"});
-        res.write(`<h2>데이터 베이스 연결 실패</h2>`);
-        res.end();
-    }
-});
-
 
 // 라우터 객체 등록
 app.use('/', router);
+
+/*router.route('/process/addUser').post(userRoute.signUp);
+router.route('/process/login').post(userRoute.login);
+router.route('/process/userList').post(userRoute.userList);*/
 
 // 404 오류 페이지 설정
 const errorHandler = expressErrorHandler({
@@ -339,7 +107,6 @@ const errorHandler = expressErrorHandler({
 
 app.use(expressErrorHandler.httpError(404, 'Page Not Found'));
 app.use(errorHandler);
-
 
 
 // 서버 시작
