+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "do-it-node",
-  "version": "1.2.0",
+  "version": "1.3.0",
   "private": true,
   "scripts": {
     "dev": "NODE_ENV=development nodemon app.js",
